Add constant dir to store constants.
