+++ backend-app/constants/Actions.js
@@ -1,0 +1,13 @@
+/**
+ * @enum {string}
+ */
+const Actions = {
+  DELETE_USER: 'DELETE_USER',
+  BAN_USER: 'BAN_USER',
+  UPDATE_USER: 'UPDATE_USER',
+  UPDATE_CALANDER: 'UPDATE_CALANDER',
+  REMOVE_SUPER_ADMIN: 'REMOVE_SUPER_ADMIN',
+};
+Object.freeze(Actions);
+
+module.exports = Actions;
